requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.3.0
appdirs==1.4.4
asyncpg==0.29.0
attrs==23.2.0
Automat==22.10.0
bcrypt==4.0.1
certifi==2024.2.2
click==8.1.7
configdict==2.10.2
constantly==23.10.4
coverage==7.5.3
decorator==5.1.1
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi_camelcase==2.0.0
fuzzywuzzy==0.18.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
iniconfig==2.0.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mirakuru==2.5.2
mypy==1.10.0
mypy-extensions==1.0.0
orjson==3.10.3
outcome==1.3.0.post0
packaging==24.0
passlib==1.7.4
pluggy==1.5.0
port-for==0.7.2
psutil==5.9.8
psycopg==3.1.19
psycopg2==2.9.9
pydantic==2.7.1
pydantic-async-validation==0.2.0
pydantic_core==2.18.2
Pygments==2.18.0
pyhumps==3.8.0
PyJWT==2.8.0
pytest==8.2.1
pytest-asyncio==0.23.7
pytest-postgresql==6.0.0
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
PyYAML==6.0.1
rich==13.7.1
setuptools==70.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.30
SQLAlchemy-Utc==0.14.0
starlette==0.37.2
tabulate==0.9.0
text-unidecode==1.3
tornado==6.4
trio==0.25.1
typer==0.12.3
typing_extensions==4.12.0
ujson==5.10.0
uvicorn==0.29.0
uvloop==0.19.0
watchdog==4.0.1
watchfiles==0.21.0
websockets==12.0
zope.interface==6.4.post2
